+++ CHANGES.md
@@ -21,6 +21,7 @@ v3.0.3 - YYYY-MM-DD
 - Fixed potential buffer overrun in `cupsDoAuthentication` (Issue #156)
 - Fixed potential buffer overrun in `cupsSignCredentialsRequest`.
 - Fixed X.509 name checks using OpenSSL.
+- Fixed renewal of expired self-signed certificates.
 
 
 v3.0.2 - 2026-06-05